+++ docs/changes.rst
@@ -8,6 +8,8 @@ Change log
   instances isn't a valid use of Django but, again, django-debug-toolbar
   shouldn't crash.
 * Added pyflame (for flame graphs) to the list of third-party panels.
+* Fixed the cache panel to correctly count cache misses from the get_many()
+  cache method.
 
 3.4.0 (2022-05-03)
 ------------------